CDash reporter is constructed with a configuration object
