+++ lib/spack/spack/reporters/junit.py
@@ -2,9 +2,7 @@
 # Spack Project Developers. See the top-level COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
-
 import os.path
-import posixpath
 
 import spack.tengine
